cloud/requirements.txt
Flask==2.1.0
Werkzeug==2.1.0
gunicorn==20.1.0
pypdf==3.1.0